+++ wh_application/src/WebHemi/Acl/AclMiddleware.php
@@ -1,27 +1,39 @@
 <?php
 
+/**
+ *
+ * WebHemi
+ *
+ * PHP version 5.6
+ *
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://webhemi.gixx-web.com/license/new-bsd
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@gixx-web.com so we can send you a copy immediately.
+ *
+ * @author    Gabor Ivan <gixx@gixx-web.com>
+ * @copyright 2012 - 2016 Gixx-web (http://www.gixx-web.com)
+ * @license   http://webhemi.gixx-web.com/license/new-bsd   New BSD License
+ * @link      http://www.gixx-web.com
+ *
+ */
 namespace WebHemi\Acl;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Zend\Expressive\Router\RouterInterface;
 
+/**
+ * Class AclMiddleware
+ * @package WebHemi\Acl
+ */
 class AclMiddleware
 {
-    /**
-     * @var RouterInterface
-     */
-    private $router;
-
-    /**
-     * AclMiddleware constructor.
-     * @param RouterInterface $router
-     */
-    public function __construct(RouterInterface $router)
-    {
-        $this->router = $router;
-    }
-
     /**
      * @param ServerRequestInterface $request
      * @param ResponseInterface $response
@@ -32,13 +44,8 @@ class AclMiddleware
      */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
     {
-        echo 'Acl<br>';
-
-        $params = $request->getQueryParams();
-
-        if (isset($params['acl'])) {
-            throw new \Exception('Forbidden', 403);
-        }
+        // TODO: build ACL graph, check accessibility, throw an error when forbidden
+        //throw new \Exception('Forbidden', 403);
 
         return $next($request, $response);
     }