+++ wh_application/src/WebHemi/Application/ApplicationMiddleware.php
@@ -1,16 +1,39 @@
 <?php
-
+/**
+ *
+ * WebHemi
+ *
+ * PHP version 5.6
+ *
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://webhemi.gixx-web.com/license/new-bsd
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@gixx-web.com so we can send you a copy immediately.
+ *
+ * @author    Gabor Ivan <gixx@gixx-web.com>
+ * @copyright 2012 - 2016 Gixx-web (http://www.gixx-web.com)
+ * @license   http://webhemi.gixx-web.com/license/new-bsd   New BSD License
+ * @link      http://www.gixx-web.com
+ *
+ */
 
 namespace WebHemi\Application;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Zend\Expressive\Router;
-use Zend\Expressive\Template;
-use Zend\Stdlib\ArrayUtils;
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\ServiceManager;
 
+/**
+ * Class ApplicationMiddleware
+ * @package WebHemi\Application
+ */
 class ApplicationMiddleware
 {
     /**
@@ -35,7 +58,7 @@ class ApplicationMiddleware
      */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
     {
-        echo 'ApplicationMiddleware (wh)<br>';
+        // todo find a usage for this middleware: before the router
 
         return $next($request, $response);
     }