+++ wh_application/src/WebHemi/Application/ApplicationMiddlewareFactory.php
@@ -1,14 +1,36 @@
 <?php
 /**
- * @see       http://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
+ *
+ * WebHemi
+ *
+ * PHP version 5.6
+ *
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://webhemi.gixx-web.com/license/new-bsd
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@gixx-web.com so we can send you a copy immediately.
+ *
+ * @author    Gabor Ivan <gixx@gixx-web.com>
+ * @copyright 2012 - 2016 Gixx-web (http://www.gixx-web.com)
+ * @license   http://webhemi.gixx-web.com/license/new-bsd   New BSD License
+ * @link      http://www.gixx-web.com
+ *
  */
 
 namespace WebHemi\Application;
 
 use Interop\Container\ContainerInterface;
 
+/**
+ * Class ApplicationMiddlewareFactory
+ * @package WebHemi\Application
+ */
 class ApplicationMiddlewareFactory
 {
     public function __invoke(ContainerInterface $container)